Modifications to the 08_ESLbyGWL jupyter notebook to make it run by Mario.
Modifications to the 08_ESLbyGWL jupyter notebook to make it run by Mario. The input data paths aere sent to archive so that all the path with input data has to be changed. It is not possible anymore to write data in the archive p drive so the output paths have to be changed.